backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
nuscenes-devkit
numpy==1.26.4
pydantic==2.9.0
pydantic-settings==2.5.2
